Add Kordal's July 2013 update for Arduino Due compatibility
The 'int' in the union for the x,y,z changed into int16_t to be compatible with Arduino Due.

+++ Gy521_send_serial/Gy521_send_serial.ino
@@ -651,13 +651,13 @@ typedef union accel_t_gyro_union
   } reg;
   struct 
   {
-    int x_accel;
-    int y_accel;
-    int z_accel;
-    int temperature;
-    int x_gyro;
-    int y_gyro;
-    int z_gyro;
+    int16_t x_accel;
+    int16_t y_accel;
+    int16_t z_accel;
+    int16_t temperature;
+    int16_t x_gyro;
+    int16_t y_gyro;
+    int16_t z_gyro;
   } value;
 };
